Update Medication Log Service Refactor
