+++ Cohort1/cohortProject/src/main/java/com/cohortE/cohortProject/service/DailyLogService.java
@@ -1,0 +1,5 @@
+package com.cohortE.cohortProject.service;
+
+public interface DailyLogService {
+    void addNewDailyMedicationLog();
+}
