+++ Cohort1/cohortProject/src/main/java/com/cohortE/cohortProject/service/MedicationLogService.java
@@ -8,5 +8,4 @@ public interface MedicationLogService {
 
     public List<MedicationLogDto> getDailyUsersMedicationLogs();
 
-    void addNewDailyMedicationLog();
 }
